Split JavaDebugger into 3 classes: -JavaDebugger contains all debugging specifics -Probes contains helpers for the probes -DebugUtils contains various helper functions
Also Group functions in JavaDebugger into logical groups
This should make it easier to find your way around the class
