+++ OODebug/src/debugger/JavaDebugger.h
@@ -29,27 +29,16 @@
 #include "../oodebug_api.h"
 #include "jdwp/DebugConnector.h"
 #include "../overlays/PlotOverlay.h"
-
-namespace Export {
-	class TextToNodeMap;
-}
+#include "Probes.h"
+#include "DebugUtils.h"
 
 namespace Model {
 	class Node;
 	class TreeManager;
 }
 
-namespace OOModel {
-	class Class;
-	class Expression;
-	class Method;
-	class VariableDeclaration;
-	class StatementItem;
-}
-
 namespace Visualization {
 	class Item;
-	class IconOverlay;
 }
 
 namespace OOVisualization {
@@ -62,16 +51,9 @@ namespace Interaction {
 
 namespace OODebug {
 
-class Location;
 class BreakpointEvent;
 class SingleStepEvent;
-class VariableDetails;
-class Value;
-
-class JavaDebugger;
-
 struct VariableObserver;
-
 struct EnvisionVariable;
 
 class OODEBUG_API JavaDebugger
@@ -91,54 +73,39 @@ class OODEBUG_API JavaDebugger
 		Interaction::CommandResult* probe(OOVisualization::VStatementItemList* itemList,
 													 const QStringList& arguments, int itemIndex);
 
-		using ValueCalculator = std::function<double(QList<double>)>;
-		using ValueHandler = std::function<void(JavaDebugger*, Values, QList<ValueCalculator>, Model::Node*)>;
+		using ValueHandler = std::function<void(JavaDebugger*, Values, QList<Probes::ValueCalculator>, Model::Node*)>;
 
 	private:
-		using ValueOperator = std::function<double(double, double)>;
 		JavaDebugger();
-		void addBreakpointOverlay(Visualization::Item* target);
-		QString jvmSignatureFor(OOModel::Class* theClass);
-		/**
-		 * Returns a String with all containing module names split by \a delimiter in front of the \a theClass name.
-		 */
-		QString fullNameFor(OOModel::Class* theClass, QChar delimiter);
 
 		bool isParentClassLoaded(Model::Node* node);
 		void breaktAtParentClassLoad(Model::Node* node);
-		Location nodeToLocation(Model::Node* node);
-		Model::Node* locationToNode(Location location, bool& isClosingBracket);
 
-		void resume();
 		void trySetBreakpoints();
 		void removeBreakpointAt(Model::Node* node);
+
+		void resume();
+
+		// Event handlers
 		void handleVMStart(Event);
 		void handleClassPrepare(Event);
 		void handleBreakpoint(BreakpointEvent breakpointEvent);
 		void handleSingleStep(SingleStepEvent singleStep);
 
-		Protocol::Tag typeOfVariable(OOModel::Method* containingMethod, VariableDetails variable);
-		Protocol::Tag typeExpressionToTag(OOModel::Expression* e);
-
-		static OOModel::VariableDeclaration* variableDeclarationFromStatement(OOModel::StatementItem* statement,
-																									 QString variableName = "");
-
-		void toggleLineHighlight(Visualization::Item* item, bool highlight, bool closingBracket = false);
-
+		// Probe helpers
 		QPair<PlotOverlay::PlotType, ValueHandler> defaultPlotTypeAndValueHandlerFor
 			(QList<EnvisionVariable> variableInfos);
-		QPair<QList<ValueCalculator>, QStringList> parseProbeArguments(QStringList arguments);
-		ValueOperator operatorFromString(QString operatorString);
-		void handleValues(Values values, QList<ValueCalculator> valueCalculators, Model::Node* target);
-		void handleArray(Values values, QList<ValueCalculator> valueCalculators, Model::Node* target);
-		double doubleFromValue(Value v);
+		void handleValues(Values values, QList<Probes::ValueCalculator> valueCalculators, Model::Node* target);
+		void handleArray(Values values, QList<Probes::ValueCalculator> valueCalculators, Model::Node* target);
+
+		// Overlay functions
+		void addBreakpointOverlay(Visualization::Item* target);
+		void toggleLineHighlight(Visualization::Item* item, bool highlight, bool closingBracket = false);
 		PlotOverlay* plotOverlayOfNode(Model::Node* node);
-
-		bool hasPrimitiveValueType(Protocol::Tag tag);
-
 		void removeObserverOverlaysAt(Model::Node* node, Visualization::Item* nodeVisualization);
 
 		DebugConnector debugConnector_;
+		DebugUtils utils_{&debugConnector_};
 
 		// For each class we should only break at loading once, otherwise we get multiple events.
 		QSet<Model::Node*> breakOnLoadClasses_;
@@ -149,8 +116,6 @@ class OODEBUG_API JavaDebugger
 		qint64 currentThreadId_{};
 
 		QMultiHash<Model::Node*, std::shared_ptr<VariableObserver>> nodeObservedBy_;
-
-		std::shared_ptr<Export::TextToNodeMap> exportMap_;
 
 		static const QString BREAKPOINT_OVERLAY_GROUP;
 		static const QString PLOT_OVERLAY_GROUP;
